requirements.txt
numpy<2
whisperx>=3.1.1
faster-whisper>=1.0.3
ctranslate2>=4.3.0
torch>=2.1.0
jiwer>=3.0.4
librosa>=0.10.2.post1
soundfile>=0.12.1
pyloudnorm>=0.1.1
openai>=1.30.0
python-dotenv>=1.0.1   # if you decide to use a .env